pyproject.toml
[project]
name = "file-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.5.0",
    "python-dotenv>=1.0.0",
]
